refactor: move LazyRead into KeyValueAccess
* remove unused KeyValueAccessLazyReadData
* make KeyValueAccessReadData package-private
